requirements.txt
setuptools>=68.0.0
wheel>=0.40.0
flask==2.0.1
tensorflow-cpu==2.16.1
numpy>=1.24.3
opencv-python-headless==4.8.1.78
gunicorn==20.1.0
Pillow==10.0.0
python-dotenv==0.19.0
werkzeug==2.0.1
scikit-learn==1.3.0 